+++ src/components/Blobs/BlobPortrait.svelte
